backend/requirements.txt
# Core Dependencies
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Google Gemini AI
google-generativeai==0.3.2

# HTTP Requests
requests==2.31.0

# Date/Time handling
python-dateutil==2.8.2

# Async support (optional)
aiohttp==3.9.1

# Data validation
validators==0.22.0

# IP address handling
ipaddress

# DNS lookups
dnspython==2.4.2
